+++ link_accounts_client.js
@@ -25,6 +25,7 @@ import './community-services/venmo'
 import './community-services/vk'
 import './community-services/wechat'
 import './community-services/line'
+import './community-services/office365'
 
 Accounts.oauth.tryLinkAfterPopupClosed = function (credentialToken, callback) {
   const credentialSecret = OAuth._retrieveCredentialSecret(credentialToken)